Add configurable search for Include Article
Before RT 5.0 the search box to include an article (only shown when the
number of articles exceeds a configurable limit, default 50) searched on
multiple fields, but in the update to 5.0 the search was restricted to
only the Article Name.

This expands on the previous functionality, patterning it after User
search by making the list of matched fields configurable.