Allow DefualtCatalog to be unset in Web Interface
Previously it was listed as needing to be set, even though the default
is unset, so the selectbox for it in the Assets subsection of the System
Configuration page was created without a "(no value)" option.  The
result of this is that any change made in that subsection would set the
DefaultCatalog to whichever catalog happened to be first in the list.

It also prevented it from processing a "Reset" as the handling of the
form data wouldn't accept having nothing selected.