Allow HTML signatures
If the user has their preferences set so that they use the WYSIWYG
editor for ticket updates then the signature textarea in their
Preferences will also be a WYSIWYG editor, allowing them to create an
HTML signature.

Signatures are automatically upconverted from plain text to HTML either
when the preferences are updated or each time corresponding on a ticket.

HTML signatures are automatically downconverted when they are added to a
plain text message.