Check email of custom role members on ticket create
Previously on ticket create the email adddresses of custom role members
were not checked to verify that they weren't known to be used by RT to
receive email.  This would allow for the creation of mail loops from RT
to itself.

Now custom roles are checked and processed the same as the built-in
roles Requestor, Cc, and AdminCc.