Allow browser spellchecker to work in CKEditor windows
This setting was lost from RT 4.4 when we removed other
CKEditor settings that were explicitly setting values that
were already the default in CKEditor.

The documentation and naming of this option are a bit misleading,
but this change allows the browser spell checker to underline
mis-spelled words and right-click shows suggestions.

Fixes: I#37323