Set X-Content-Type-Options to nosniff to tell browser not to change content-type
This addresses CVE-2022-25802.

When browser encounters a content-type it doesn't recognize, it can
"sniff" the content and choose one suitable for the content, which is
risky as it could execute the content(e.g. as HTML with js) that is from
untrusted external source.

Setting X-Content-Type-Options to "nosniff" disables browser's "sniff"
behavior. By setting it here, it not only fixes the issue above, but
also potentially provides better CORB-protection for various content
types.

See also https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Content-Type-Options

There is a config option to disable this if necessary.