Search attributes with extra limits on a clean cloned search builder object
"Attributes" method returns a cached search builder object, so if we
want to add extra limits, we shouldn't touch the original cached object,
otherwise the query wouldn't be right, e.g.

    my $attributes = $object->Attributes;
    $attributes->Limit( FIELD => 'Name', VALUE => 'Dashboard' );
    $attributes->Limit( FIELD => 'Description', VALUE => 'foo' );

Later,

    my $attributes = $object->Attributes;
    $attributes->Limit( FIELD => 'Name', VALUE => 'Dashboard' );
    $attributes->Limit( FIELD => 'Description', VALUE => 'bar' );

For $RT::System, as we don't re-initialize it much, the 2 snippets above
share the same $object, this would unexpectedly get global dashboards
with Description "foo" *or* "bar".

Note that it's not enough to create a brand new object every time in
_LoadObject, as _UpdateObject uses the same object in a loop, where
attribute searches occur.