Put all dependencies of current object to the head of stack
Previously we put objects depending on current object to the end of
stack, which could result in a really big stack and was not efficient.

Assuming a much simplified case: 1000 tickets(each has 10 transactions).
When 500 tickets got processed, and the stack was like(Page=100):

    tickets(contains 501-1000)
    transactions of ticket 1
    transactions of ticket 2
    transactions of ticket 3
    ...
    transactions of ticket 500

As ticket transactions were put after *all* tickets, the stack size
could be huge. Things are even worse in real world as transactions have
attachments, attributes, etc, which would also be put to the end of
stack and wouldn't be processed until all transactions are serialized.

With this commit, since transactions of processed tickets are processed
right after their related tickets, the stack is like:

    tickets(contains 501-1000)

which is much more memory efficient.