Improve recognition of urlified subject tags
This change improves RT's handling of subject tags that have been
modified by an external mail client to include a http:// prefix (e.g.
because the mail client believes it looks like the name of a possible
web server).

A previous change allowed RT to recognize the modified tags enough to
not create new tickets, but in some places they were still being
detected as tags from another RT system, resulting in the ticket subject
being modified.

With this change, RT now completely treats the modified tag the same as
it treats the unmodified tag, which will preserve the ticket subject
unchanged.