Improve checking of CustomFieldValue SortOrder
Previously there was no proactive validation on the SortOrder for
CustmField values.  Some databases are more picky than others and throw
errors for non-integer values.  In one case the DBD error message was
even returned to the user.

Now SortOrder is validated to be either empty or contain an integer
value (negative values are fine), plus a (more) user-friendly and
localisable error message is generated.