Format html custom field changes like other text ones
This is to show the diff, which was implemented in 290067f80a.

Here we call the corresponding ShowCustomField... components to render
content, which is more precise(as Text/Wikitext/HTML are treated a bit
differently). As these components append a newline to the output, we
need to use a more tolerant regex in tests.