Remove submit from theme Try button
As part of the updates to bootstrap in e3bed571, the Try
button on the Theme editor was converted to a submit.
The result is that instead of just reloading the test CSS
in the current browser via javascript, it saves the update
to the database and sets it for all users.

Remove the submit and restore it to just a button.

This also makes the Reset button useful again as it can
remove the settings applied with Try.