Make simple search result refresh always function
There is a circumstance in which simple search results cannot be
refreshed. Doing a simple search and selecting a refresh interval always
works, but the next simple search performed without changing the
refresh interval will produce the error "There was an error parsing your
search query..." This prevents that error from happening.

Simple search uses share/html/Search/Results.html to
display its results. When a refresh interval is selected, it is
submitted as a form parameter to /Search/Results.html, and a CSRF token
is generated for that request path. When the next simple search is
performed, the search term is submitted to /Search/Simple.html, and a
CSRF token is generated for that path. When the results auto-refresh,
however, the path /Search/Results.html is what is refreshed, and that
path doesn't match the path for the CSRF token, causing the error above.

There is nothing different about the simple search results page that
distinguishes it from the results provided by Search Builder, so rather
than compose that Mason template into Search/Simple.html, redirect to
Search/Results.html like Search Builder does currently.