Fix HTML custom fields getting 'text/html' as a value
Commit c9d8c506ca added the HTML custom field type.  One of the
mechanics for properly displaying and decoding the value for these
fields is having a '-ValuesType' hidden input associated with the CF.

The functions
RT::Interface::Web::ProcessObjectCustomFieldUpdatesForCreate and
RT::Transaction::UpdateCustomFields did not know to ignore this in the
args list and were therefore processing it as if it were a '-Value'
parameter, causing a CF that wasn't otherwise given a value to get the
value 'text/html'.

Other object types with custom fields use
_ProcessObjectCustomFieldUpdates which did know to ignore the
'ValuesType' parameter.