Exempt some format strings from HTML::Gumbo structure check
6c0cbbbd introduced an extra structure check for HTML aimed
at allowing well-formed tables in email to be displayed
in ticket history.

RT also sends other types of HTML through the scrubber and
HTML::Gumbo undoes some of the escaping applied by EscapeHTML,
which is necessary for properly displaying apostrophes in
column titles. This causes the Saved Search portlet titles
to display incorrectly for titles with apostrophes.

Add an optional parameter to skip the structure check for
these cases where structure isn't an issue and the extra
parse causes problems when modifying the content.