Fix the menu drift issue when clicking repositioned submenus
jquery.supposition.js is to tweak submenu positions if they overflow
screen in original positions. E.g. to lift the Scrips submenu from:

    Admin
    Users
    Groups
    Queues
    ...
    Global  Scrips
            Conditions
            Actions
            Templates
            ...
            Topics

to

    Admin
    Users    Scrips
    Groups   Conditions
    Queues   Actions
    ...      Templates
    Global   ...
             Topics

Previously if you clicked one link in a repositioned submenu, the
submenu would be repositioned(again) back to the orignal position, which
we call a "drift" here. It is acutally a tricky precision issue, e.g.

    var expandUp = (offsetTop + menuHeight > baseline);
    if (expandUp) {
        $u.css('margin-top', baseline - (menuHeight + offsetTop));
    }

Even if "margin-top" is set correctly, "expandUp" could still be true
next time as "offsetTop + menuHeight - baseline" could be non-zero like
0.003, which would then set "margin-top" back to nearly 0, i.e. its
original position. This commit fixes this issue by allowing a bit
tolerance.