Add SameSite to cookies from WebSameSiteCookies
This change adds the SameSite option to RT's session cookies, with a
default value of 'Lax', providing a reasonable amount of protection
against CSRF attacks while still allowing most integrations of RT with
other systems.

It also provides a config option, WebSameSiteCookies to allow the value
to be changed from the default.  Current standards allow values of
'Strict', 'Lax', and 'None'.  Any value other than these will cause a
warning to be logged.  Current standards also require the Secure option
to be set on the cookie when using the 'None' value, so a warning is
logged if 'None' is specified and WebSecureCookies isn't set.