Update default value for WebSecureCookie
Previously $WebSecureCookie defaulted to '0', so RT did not set the
'Secure' attribute on the session cookies.

This changes the default to '1' so that session cookies are flagged as
'Secure' by default.  This will generally cause browsers to require an
SSL connection to the server and may also change how the cookies are
cached.