Remove state criteria for invalid utf8 error warnings
MariaDB has assigned state code 22007 to the 1366 encoding error, so it
no longer matches the error check. This causes the indexer to stop
processing rather than reporting the encoding problem and continuing.
For installed and running RTs, a record with incorrect coding will also
cause the indexer in cron to stop at that record.

Remove the state code check to allow this error to still be recognized
and skipped for MySQL and MariaDB. The error code itself should be
specific enough to handle just the encoding error we seek to skip here.