Allow empty search ResultPage parameter in URL
RT could generate URLs with empty ResultPage(e.g. when building menus),
which doesn't hurt as it falls back to "/Search/Result.html".

In ba3a82144d we added validation of ResultPage, to avoid redirecting to
arbitrary pages for security reasons. As empty ResultPage falls back
safely, allowing it is totally fine.

This is mainly to get rid of the following warning:

    ResultPage  is not whitelisted, ignoring