Use $SkipProcessing for ticket updates only
The $SkipProcessing is used to short-circuit the following ticket
updates in some cases like invalid custom field values are submitted.

No matter if $SkipProcessing is true or false, we shall always check
"ShowTicket" right and the "MarkAsSeen" can be processed too as it's
more like an update to current user instead of the ticket.

Note that it's not quite a security issue: even without the "ShowTicket"
check, ticket fields wouldn't show up on display page anyway, in which
case the page would look like a ticket without any data in it.