Unify the BeforeActionList Callback API
Two instances of the BeforeActionList Callback were missing the Actions
parameter.

This change adds the parameter on the calling side.  Implementations of
the callback will be unaffected by the addition of the new parameter,
but they now have the option of modifying the list of user-reported
messages in /Ticket/Reminders.html and /Ticket/Update.html the same as
they already can in /Ticket/ModifyLinks.html, /Ticket/ModifyAll.html,
/Ticket/ModifyDates.html, /Ticket/Modify.html, /Ticket/Display.html,
/Ticket/ModifyPeople.html, and /Ticket/Forward.html.