Provided calculated count as a callback parameter
Generating count requires running the query and this has
already been done when the callbacks run. Provide the
count in case it can be useful since the work has already
been done. This is mostly likely to be useful for callbacks
that have a different behavior if no tickets are found.