Convert ticket link graph generator to GraphViz2
The GraphViz perl module is deprecated, so convert to
GraphViz2.

As part of this update, the graphs on /Ticket/Graphs/index.html are now
displayed as SVG with links instead of as PNG with a client-side
image map. This change is to align with the interface for image maps
in GraphViz2, which is a bit different from GraphViz. The image map
and graph are both generated as part of the "run" and output together.
The GraphViz2 documentation also notes the SVG image map is more light-weight.