Fix delimiter during upgrade in recipient list
In older RT, the dashboard recipient list supported being delimited
by semicolon or space.  The 4.4.1 upgrade content step migrates the
list, but on newer RTs, fails to split the list correctly.

This commit updates the 4.4.1 upgrade content step to correct the
semicolon and space delimited list to then be correctly split and
migrated.