Add checking to verify attr_match_list maps
To allow updating user attributes in RT with concatenated
attributes from LDAP, the ExternalSettings attr_map configuration
supports defining multiple LDAP attributes as an arrayref.
RealName in this example:

 'attr_map' => {
     'Name' => 'uid',
     'EmailAddress' => 'mail',
     'RealName' => [ 'givenName', 'sn' ],
 },

However, if an entry in attr_match_list maps to an arrayref in
attr_map, Net::LDAP::Filter will not build the search filter,
and will not return a filter object, resulting in a failed
response from the LDAP search.

 'attr_match_list' => [ 'Name', 'EmailAddress', 'RealName' ],

This commit adds checking to verify entries defined in
attr_match_list don't map to an arrayref in attr_map.  If found,
the LDAP search using that attr_match_list attribute is skipped
and the error logged.