Improve "forbidden" for REST2 record endpoints with more right checks.
Previously returned "not found" if the corresponding record doesn't
exist, even when user doesn't have proper rights granted, which led to
infoleakage.

This commit tweaks the logic to check "CurrentUserCanCreate" if the
record doesn't exist, which eliminates the leakage. Besides that, it
checks "CurrentUserCanModify" instead on object update, which is more
appopriate than "CurrentUserCanSee".

There is a couple of right changes for "/customfield/" and "/group"
endpoints: previously modifying them required "SeeCustomField" and
"SeeGroup", respectively, which was not quite necessary as it's really
odd to grant users "Admin..." but not "See..." rights. This commit
dropped the "SeeCustomField"/"SeeGroup" check, to be consistent with
other objects.