Add RT::Scrip::DataCache
This change adds the DataCache accessor method to RT::Scrip objects for
use in safely passing information from one phase of scrip processing to
others, e.g. from Condtion code to a template.

Each Scrip object is created with a new empty hashref for its DataCache.
This is a transient cache and no data is preserved when the object is
destroyed.  This cache is not used for any other purpose so that there
is no concern about conflict with current or future RT Scrip attributes
or methods.