Specify equals comparison for one-time email lookup performance
The current EmailInput code queries the user autocomplete helper
with no operator, so it uses the default "UserSearchFields" which
searches on EmailAddress, Name, and RealName, all with LIKE
searches, which can be slow.

This component is an EmailInput, so values provided as options
should be emails. Set the operator to = for a faster search.
The user autocomplete defaults to searching for only EmailAddress.