Remove id from search save args
If id is listed in the search save args then the search saves the ticket id
for the ticket being viewed when the search is saved. If the search is loaded
for a different ticket it shows the graph for the original ticket and not the
one currently being viewed.