Avoid adding duplicated criteria to queries generated in JSChart
For a bar chart grouped by status, each bar represents one status, and
clicking it redirects to search result page that shows tickets with that
status only. This is achieved by adding the extra status criterion to
the original query, e.g. TicketSQL is changed from:

    Queue = 'General'

to

    ( Queue = 'General' ) AND ( (Status = 'new') )

If you go to Chart page and click again, previously we unconditionally
added the extra status criterion, so the TicketSQL would be

    ( Queue = 'General' ) AND ( (Status = 'new') ) AND ( (Status = 'new') )

which is unnecessary. This commit checks the queries beforehand to avoid
the duplication.