Support loading users via user custom fields
External auth allows admins to provide a list of attributes
for users to use in the username box on login, like 'Name'
or 'EmailAddress' and it will try both. 566b8fa1 added support
for using UserCFs in mappings, but not as login attributes.

This commit adds support for loading users with UserCFs, so
configurations like this can now be used:

    'attr_match_list' => [
        'UserCF.Employee ID',
        'EmailAddress',
    ],