Fix broken Pod in rt-validator
The Pod for check_integrity appears between the table of contents and
NAME directive in the Pod when rendered to HTML (with the link to the
function appearing first in the TOC), and rendering to text is similarly
broken.

As this is a utility rather than a library, and check_integrity is not
used outside of this utility, these docs are best represented as
comments instead.