Disable jump to page form by default in CollectionList
Some pages put CollectionList inside an existing form
and HTML doesn't like nested forms. In these cases,
adding the jump to page form can cause the existing form
to stop working, like on Bulk Update.

Make the jump to page form optional and default it to off
since it is new and may not be appropriate for all locations
where CollectionList is used.

Enable for ticket search results, which is the primary use
case for adding it.