Avoid creating duplicated custom fields from initialdata
Custom fields can have duplicated names, which makes sense considering
they can be applied to different objects. On the other hand, it's weird
and usually unintentional to apply multiple custom fields with the same
name to the same object.

This commit detects the latter case and refuses to create duplications
like that.