Always show a Logout link in the menu
The updates in 4e46f947 are intended to send all logout requests
through Logout.html first, even if another logout link is
configured. This allows us to remove the user session and write
a log message, even if we then redirect to an external logout
endpoint, as with SAML auth. Since we will always do this, we should
always show the Logout link.

Remove the code to hide the link in some cases.