Add Web External Auth logout Logic to Logout.html
We moved the URL redirection logic from MenuBuilder to the Logout.html.
This is because now we always redirect first to our Logout.html page
for generating appropriate logout messages.  Only after that the user
is redirected to the external logout URL.
Also added new logs to warn RT admin if the external LogoutURL is not
set.