Replace UNIVERSAL::require with Module::Runtime::require_module
Created a new library RT::StaticUtil which contains a function which
wraps Module::Runtime::require_module.  This is required because
Module::Runtime::require_module dies on error, and we need to catch
the error and return it to the caller.
Updated all the places that were using UNIVERSAL::require to use
RT::StaticUtil::RequireModule instead.