Replace fontawesome funnel icon with bootstrap version
We wanted a version of the funnel icon with just the outline
and no fill, and that icon is not in the fontawesome free
collection, so switch to the MIT licensed bootstrap version
which provides both the solid and outline funnels.

The color is set directly in the svg files by changing "fill"
from the default bootstrap version.

This update also makes the icon a bit smaller.