Calculate search filter modal content's max-height accurately
Previously we only set the whole modal's max-height to 80%, but since
its descendant div.modal-content's max-height was set based on
viewport("calc(100vh - 3.5rem)") in bootstrap, its height could be
bigger than its ancestor, in which case a scrollbar would show up.

This commit set div.modal-content's max-height based on viewport and the
position of the modal to make sure it's 100% viewable. With it, there is
no need to set div.modal's max-height(80%) any more.