Do not mark owner search-filter modal scrollable
Class "modal-dialog-scrollable" makes underlying ".modal-body" scroll if
needed(overflow-y:auto), which is not suitable for Owner as its original
height is quite small and not enough for expanded bootstrap select
options.

This commit gets rid of the "modal-dialog-scrollable" class so the
bootstrap select options can overflow across the whole modal height,
instead of just the small "modal-body".