Improve adding of action links to search results
This is streamlining existing functionality; previously, you'd need to
select a column, a link, and maybe a column title to add an action link
to search results. The user would end up with either a column title they
picked or the name of the field, and the clickable action would be the
value of that field. The user could then make changes to the format to
replace the header with a space (therefore removing the header) and
replacing the field value with the name of the action to take.

This commit assumes that users prefer the latter as the default
functionality, and treats the action link as other columns. Users can
then edit the format if they want to revert to the original format (or
something different, of course).