Render menu link icon as inline svg rather than CSS class
When loading search results, the page menu would load
without the link icon, all search results would load,
then the JS would run and the link icon would jump into
the menu. At the top of the screen, it was very
noticeable.

Convert to inline svg to load when the page loads and
still retain the ability to style the icon. The
new link svg is from Bootstrap icons.