Hide tooltips everywhere on click
When clicking on the filter icon on ticket search results,
the tooltip remained on the screen, sometimes blocking
the modal. Add JS to hide tooltips everywhere on click
and remove some previous individual fixes for the same
issue.

Keep the dispose for bookmarks as the "Remove" tooltip
persists and is visible behind "Add" when toggling
back and forth if dispose isn't called to clean it up.