Don't highlight "RT for" as the active menu
The "RT for example.com" menu entry has the same link as
the Home menu which causes the "Logged in as" menu to
be shown in bold as the active location along with the
Home menu. Showing the active menu is an aid to users
to show them where the current page is in the menu, but
in this case showing it twice is misleading. Since the
Home menu is the primary way to load the root RT page,
don't highlight the "RT for" menu as active when on the
home page.