Remove trailing artefacts before adding query part
Old code could add query part, eg '?foo=bar' to $base_path, so
removing trailing 'index.html' and '/' would never work.

So remove trailing parts first and only after that add query string
if it's not empty.

It only influence whether child is marked as active or not.

It is hard to see if it affects production code or not, but new
code is more correct than the old one.