Get the Stylesheet of the called user object instead of its CurrentUser
It's apparently wrong that $user->Stylesheet returns the
"WebDefaultStylesheet" pref of the user's CurrentUser object.

Previously it wasn't an issue as we call Stylesheet only on an
RT::CurrentUser object, of which CurrentUser method returned itself. It
wasn't true any more as we changed RT::CurrentUser::CurrentUser to
return a new object in b7b820d316 to avoid circular refs.

This commit fixes a web UI issue that updating "Theme" on user Prefs
page didn't take effect immediately, because of outdated cached
attributes of the object returned by RT::CurrentUser::CurrentUser.