Allow to set columns to their default value or NULL
There are a few columns that have default values defined or could simply
be NULL, disallowing to set them back is annoying and not quite right.

The code converting NULL to 0 was created 2 decades ago at which time
SearchBuilder didn't support to pass undef values yet. SearchBuilder
supports it since 1.64(a decade ago), now it's safe to drop the obsolete
code.