Adjust EmailInput element to use the correct autocomplete helper
The EmailInput element was using Autocomplete/Users as a fixed helper
for generating items and options. This was making Selectize remove group
values from the list of selected items when the page was loaded with
pre-filled values.

This patch makes the helper to be the same as the one informed by the
AutocompleteType parameter.