Disable buildkit to continue using the local network feature
Buildkit is a new build engine for docker that provides some
speed improvements by doing parallel builds, among many other
features. Unfortunately, it does not support the --network
argument, so when it became the default build engine in new
versions of docker, our github actions stopped working.

There is a bunch of discussion about the network option and
buildkit here and in other issues:

https://github.com/docker/buildx/issues/175

The simple answer for now is to explicitly disable buildkit
to use the classic builder for our tests.