Add default value for calls to LookupType
If upgrading from 5.0.1 or previous and the database has a custom role
then upgrade steps containing "content" would fail as the 5.0.4 code
expects the LookupType column to exist in the database.

Technically, RT::CustomRoles->RegisterRoles would die because of absent
LookupType, and as it happens quite early(during Init), we can't use
backcompat to fix it.

Adding a default value for calls to LookupType ensures if there is no
result the code defaults to the only possible value for older databases.