Fix Find Group portlet input size
The input had a fixed col-4 size, which was different from the input of
the Find User portlet. We refactored the GotoGroup element to accept a
cols parameter the same way GotoUser accepts it.

Adjusted the FindGroup element to pass the cols parameter to the
GotoGroup element and use the same number of columns as the FindUser.