Rebalance page menu when the entire page(not just dom) is ready
Previously page menu was rebalanced when dom was ready, which is not
good enough as element widths were not fully nailed yet by then.
Because of this, sometimes subject and page menu could still overlap.

This commit runs page menu rebalance code when the entire page is
ready(code for nav menu is already there), at which time we can get more
accurate element widths and thus rebalance more accurately.