Remove Dropdown render type for select custom fields
Both "Dropdown" and "Select box" are rendered with selectpicker, as they
behave exactly in the same way, there is no need to keep both.

Default render types are saved as 0 or NULL, as the previous default
"Dropdown" has been removed, "Select box" becomes the new default render
type for single-value selects, which means the data in old db is still
correct, there is no need to update database.