Migrate js code to ckeditor5
All outdated dark-mode related js is dropped, we will implement it in
CSS instead.

To make it work better with the upcoming htmx(where we need to locate
ckeditor objects via submitted textarea names), here we keep track of
ckeditor objects using corresponding textareas' name instead of id: the
former is more correct anyway since mark_changed function already assumes
so, not mentioning that we have some textareas that don't have id set.

As we are going to drop jQuery in RT 6, new code gets rid of it on
purpose.