Process ticket owner updates before message updates
In an RT configured to send notifications to some users
only when they are an owner, performing a single update
that sets a new user as the owner and sends a message
results in the new owner not getting email for the
reply/comment, because the message was processed before
the owner update change was made. This can be confusing,
especially if the message contains notes specifically
addressed to the new owner, such as the reason they are
being assigned the ticket.

Add a new Process step ProcessTicketOwnerUpdate to run
before ProcessUpdateMessage so the new owner is set
before notifications are sent. This allows the new owner
to get email for associated reply/comment.