Fix Enable checkbox behavior on Scrip Creation
In fee8a9b4f0, the previous EditScrip page was split into Create and
Modify. In the process of doing so, some elements were not properly
adapted to the Create scrip page. This led to the fact that the
"Enable" checkbox had no effect on Scrip Creation page and Scrips were
being created as enabled regardless of the checkbox state.