Add ability to copy a scrip to create new
Previously, to copy the parameters from an existing scrip to a new one,
you have to manually recreate each field.

This commit adds a "Copy" link to the Modify page for a scrip.  By
clicking the link, RT will clone the scrip and disable the new version.
It will also copy the applied objects and stages.

This also updates the Stage method of Scrip, allowing it
to receive queue_id named argument to get the Stage of
the scrip informing the queue directly and not only
the ticket.