Add "required" attribute for mandatory custom field inputs
The attribute isn't enabled for "List" style of multi-value select
custom fields: "required" means each "required" checkbox must be
checked, but we want to allow to check just some of them.

For mandatory select inputs, previously in a67e05ea49 we chose to keep
"(no value)" as the initial state to force user to manually select one
valid option. This commit implements the logic in js, and there is no
need to keep the invalid "(no value)" option any more.