Set id as the PRIMARY KEY of AttachmentsIndex for Pg
This is consistent with mysql's fulltext table. As id is the primary
key, DB automatically indexes the column which helps performance
especially for the following SQL that is called quite frequently when
indexing data:

    SELECT MAX(id) FROM AttachmentsIndex

Mainly to support customized fulltext table name, we implement the
upgrade in Perl instead of SQL. Use "indexes" instead of "content"
considering the database handle "content" uses usually doesn't have the
permission to alter tables, not mentioning that it's more like an index
change.