Add CSS definition for "pre" blocks
Bootstrap sets the "pre" font-face to "Inter" before "monospace", which
ends up using the "Inter" font for "pre" blocks.  This is not ideal for
code blocks because it is not a monospaced font and will render code
or inline tables misaligned.

This commit adds a CSS definition for "pre" blocks that sets the
font-face to "monospace" with a fallback to "Courier".