Exclude radio and checkbox inputs from duplicate check for inline edit
All input elements in a radio button or checkbox list have the same
name, so the rule excluding inputs with a duplicate name was also
missing any "checked" inputs. Using inline edit mode "always",
this resulted in radio and checkbox CFs getting unset when a form
field in another portet was updated. Excluding these from the name
duplicate check allows the form values to be cloned and submitted
correctly.