Fix use of loop variable
I was having this test fail when malware scanning is enabled with:
Use of uninitialized value $_ in concatenation (.) or string at t/web/charting.t line 14.

In perlsyn under https://perldoc.perl.org/perlsyn#Foreach-Loops it says:

The foreach keyword is actually a synonym for the for keyword, so you can use
either. If VAR is omitted, $_ is set to each value.

VAR is set (it is $n), so perhaps this usage of $_ while VAR is set
is undefined.