No need to initialize RT if it's just to check syntax
This is initially for DBD::MariaDB with which "perl -c -MRT=-init ..." would
generate an annoying message like:

    (in cleanup) Can't call method "_should_log" on an undefined value

It indicates that the DBI object was destroyed too late somehow with
DBD::MariaDB.

Test t/99-policy.t fails because of it. As -c runs all BEGIN blocks, here
we convert related code in rt-setup-fulltext-index to "use RT '-init';"
instead, which has proper handling of -c.