Split fulltext search terms for both correctness and performance
Previously for searches like:

    Content LIKE "foo" AND Content LIKE "bar"

If a ticket has "foo" saved in one attachment and "bar" in another, it
wouldn't match because these fulltext criteria needed to match on a single
attachment. What we want is to match on ticket level, i.e. as long as the
ticket has "foo" and "bar" saved in some attachments of the given ticket, it
should match.

Thus we split the query into 2 and intersect them:

    (Content LIKE "foo") INTERSECT (Content LIKE "bar")

For searches like:

    Content LIKE "foo" OR Subject LIKE "bar"

We split the query and unite them accordingly:

    (Content LIKE "foo") UNION (Subject LIKE "bar")

The latter version has much better performance as it makes use of fulltext
indexes, unfortunately the former version can not right now.

See also 8450f0a9f2