Make sure TicketType and ObjectType limits in transaction searches are AND-ed
To add TicketType='ticket' and ObjectType="RT::Ticket" preconditions,
previously we wrapped user's TxnSQL with only if it doesn't contain outer
parens, of which the regex check was not quite accurate. E.g.

    ( Type = 'Set' ) OR ( Type = 'Correspond' )

It matches the regex(qr/^\s*\(.*\)$/) but actually we still need to wrap it
so the final TxnSQL could be:

    TicketType = 'ticket' AND ObjectType = 'RT::Ticket' AND ( ( Type = 'Set' ) OR ( Type = 'Correspond' ) )

Otherwise all Set/Correspond transactions would be included in search
result.

To fix the issue above, this commit parses user's TxnSQL to wrap it as long
as the root condition is "OR".