Make sure TicketType and ObjectType limits in transaction searches are AND-ed
To add TicketType='ticket' and ObjectType="RT::Ticket" preconditions,
previously we wrapped the user's TxnSQL only if it didn't contain outer
parens, and the regex check was not quite accurate. E.g.

    ( Type = 'Set' ) OR ( Type = 'Correspond' )

It matches the regex(qr/^\s*\(.*\)$/) but actually we still need to wrap it
so the final TxnSQL will be:

    TicketType = 'ticket' AND ObjectType = 'RT::Ticket' AND ( ( Type = 'Set' ) OR ( Type = 'Correspond' ) )

Otherwise all Set/Correspond transactions would be included in the search
result.

To fix the issue above, this commit parses the query, then wraps
it if the root condition is "OR".