Fix the regex that identifies if the item is search term for autocomplete
All digital items that may or may not have type prefixes(like "asset:") are
existing values, the rest are search terms.

When we switched to LinkTargets autocomplete in 6d0b00e669, we updated the
regex to handle non-ticket link types like "asset", "articles", etc. The
regex wasn't correct as it treated "foo" in "foo bar" as an existing value,
thus if you selected the ticket from dropdown, the input would be like "foo
123" instead of "123".