Display first accessible catalog when default unset
Users were wrongly receiving a permission denied error when no default
catalog was set on their account or if the default catalog of the system
was one that they did not have access to, or even if there was not default
catalog set on the system.

This change will show the first available catalog for the user when no
default catalog is set based on the catalogs that the user has access to
create assets in.