Sync checkboxes before deciding to check/uncheck TxnSendMailToAll
For users with SimplifiedRecipients turned on, there are duplicate
"TxnSendMail" inputs in "Recipients" and "Scrips and Recipients",
respectively. In the call of "setCheckbox(this)", we automatically
check/uncheck "TxnSendMailToAll" inputs based on if all "TxnSendMail" inputs
are checked, thus we need to make sure these "TxnSendMail" inputs are synced
before that.

Previously "syncCheckboxes" and "setCheckbox(this)" were called in 2 event
triggers, which couldn't guarantee the order.